renamed incomprehensible names of methods and classes, added seperate Dtos

+++ src/main/java/com/exadel/sandbox/permission/repository/PermissionRepository.java
@@ -3,7 +3,6 @@ package com.exadel.sandbox.permission.repository;
 import com.exadel.sandbox.permission.entity.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {
-
+public interface PermissionRepository extends JpaRepository<Permission, Long> {
 }
